+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add missing jquery requirement for datatables package
 - fix bug in collection preventing image deletion in record stores
 - fix bug in number field validator
+- escape reserved words in store find methods
 
 ### Added